OOD for GUI
Created a basic OOD for the gui breaking down each piece into a class. Was having trouble running the code so will need to debug. Not displaying the GUI in the current state. Plan is to get the main part up and then add on the classes as I go.