A controller, finally...
With some pain, we now have a controller.  The flow of the software is slightly different now.  __main__ starts Tk and the controller.

The controller then starts the model (which is previously the session approach from UUTrack), then the camera, and then the view.

Within the controller we can now directly add the event tracing.

Thus, the controller now knows what the model and the view are doing.

The view currently knows the model, but strictly speaking, it shouldn't. In general, we could move the prepopulated values from the view back to the controller.  However, the view knows nothing about the controller.

The model knows nothing of the view or the GUI.