Organization, filling out controller
Decided that the interfaces to the devices best belong with the controller.  In my mind, the model is something that describes the state of the microscope.  The controller takes that information, updates the GUI, and executes the commands.  So really, the controller is a bunch of functions, which is what the devices exactly are.