Startup popup idea for config selection
Idea is to have a popup that allows us to select if we want synthetic or not so that we do not have to manually alter the yaml file which could cause issues down the line. Quality of life improvement. Solving this will also help establish some structure to how we handle popups/model changes in general