Task #2 Widgets Added and Refactored
Added some refactored code for the channel settings widgets using loops, lists etc. Will prob need some more but its reducing the code. The controller now only has one line to set the first channel active and then a loop with 4 lines to init the rest of the widgets. You can see the changes made to the populate lasers GUI and then that those groups of lines are commented out and replaced by the loop.