Moved concurrency tools to model, and identified one bug in NIDAQ
Concurrency tools will eventually be acting on the devices within the model, so I figured that is actually where it belongs.  Also found one bug in NIDAQ.py, and still need to confirm that the voltages being delivered are appropriate with an oscilloscope.

Quickly tried to put the devices in the ObjectInSubprocess concurrency tool, but received a strange attribute error that the module threading has no attribute excepthook, which it does. Curious if we have a python version specific issue there.

https://docs.python.org/3/library/threading.html