Calculate Experiment Duration
#22 - @annie-xd-wang - I have pushed a function in the channels_tab_controller that aims to calculate the duration of time for the entire experiment to run.  This would also be the framework for how we actually run the software too, except where we are simply calculating the time, we would actually be triggering commands to the hardware.

You will have to check to make sure that I am calling some of the parameters correctly.  But I tested it with some fake parameters in a jupyter notebook, and it worked. We will then need to have it send the values back to the View.

More functionality needed for the multiposition control.