Yaml refactor (#31)
* Updated configuration and experiment yaml files

All the calls to it are now broken.  Working on that next.

* Fixed paths in yaml refactor

I suspect there are more path breaks lurking, but using the --synthetic_hardware call it is now bringing up the GUI.  Available for evaluation @codeCollision4 and @annie-xd-wang.  Hopefully the experiment file that I put up is also clear.  I am available for a discussion if you would like probably around 12:30 PM my time.

* Added documentation for the remote focusing devices