Cam view (#35)
* Starting #25 Created cam view controller

* Adding Controller Call for #25

* Adding import statement for controller

* Add to image mode of execute function

#25

* Adding continuous mode to acquire execute command

This is to catch any functionality that needs to be activated when acquire is pressed and mode is 'continuous'

* Stop Acquire #25

Adding a if statement to catch anything that needs to be deactivated when stop is pressed when in continuous mode

* Pull Request Test (#33)

* Moved camera display folder for better organization

Adjusted import statements due to move

* Further Updating Imports From File Restructuring

* Added camera_view_controller working on #25

This commit shows what needs to be changed when adding a new controller for a group of widgets or a tab.

* Revert "Added camera_view_controller working on #25"

This reverts commit c7d8c494078f71a6e9321cfafdeb0417e8d1dcd8.

* Thread Pool vs ObjectInSubprocess

* Note to self

* Camera View Tab main change

Added a black image to show size and as a placeholder for the frames. Resized the canvas

* Synthetic Camera now displays an image

There is some slight frames dropped when there is a read so we will need to buffer to prevent but that can come later. For now its something.

* Adding comment about bug