Created a batch file for startup
Please change the paths to match your computer then save this file to the desktop. You can double click it and it will open a cmd and start the software.
The root should be the path to your anaconda. This will be further optimized down the line but its a start.