Starting Custom Popup #37
Had some existing code, creating an example for the acquire popup. Idea is to use the popup class and call it like you would a combobox widget. Comments try to explain. Then in the acquire popup example you just create the popup widget and add widgets to that in the class to remove the repetitive code of creating the popup window