Debugged Laser, DAQ, and snap image for the camera
Hamamatsu now takes in a single image when the mode is set to Single Acquisition. The changes in the main controller execute function need to be more elegant and possibly reorganized/refactored but the functionality is there. We also made if statements to categorize each mode for now. This can be reorganized as well, just a visual helper.

Co-Authored-By: Dax Collison <79557266+codeCollision4@users.noreply.github.com>