Setting Up #46
This also includes cleaning up some files that are no longer necessary for the popup. Also had to change some code for when we want a popup to always be present with the main window and also be able to take in input at the same time as the main window.