Remote Popup Done
Closes #46

Will just need to add a button to GUI so that it can be accessed. I added one extra parameter to the generic popup for topmost control. LabelInput class now has a function to format input widget within. I also messed around with the widget formatting so if the layout looks tidy I'll move forward with that style for now.