Update contrast_gpu.py
Trying to finish all of the operations with TensorFlow...  Getting a NaN for a entropy value.  Don't have time to look into it.  But the return on investment looks minimal.  When I switch back to a numpy array and then calculate the entropy there, it takes the same amount of time for me to get the NaN value incorrectly with tf.