Added some test code and a hoverbar class
Hover bar class example is in the center widget when running:
python validation.py

Its from stack overflow so I will need to reread the code to get more accustomed so I can customize it.

Still playing around with it. I need to propagate the code to the actual GUI but prototype semi works.