#49 #20 Validation/Error Handling
This is an example of how a form or feature will not work/save if there is an error in a specified field. For fields that need validation and data we can use the validated widgets to flag them. If there is an error then the popup in this example will not save until data input is as expected.