Camera Work
I wanted to make sure that we could make the calls to the camera for the readout direction, sensor mode, etc.  I also now calculate the readout time and the maximum frame rate.

It would be valuable if @annie-xd-wang could check the changes I made to the HamamatsuAPI file.  Now, when we call set_property_value (line 847), it calls get_property_range (line 840) to evaluate whether or not the value we are trying to set is appropriate.

In the mesoSPIM control repository, they choose whether or not they should export the values as a float or an int based upon a variable called DCAMPROP_TYPE_MASK (see line 468 here: https://github.com/mesoSPIM/mesoSPIM-control/blob/master/mesoSPIM/src/devices/cameras/hamamatsu/hamamatsu_camera.py).

I did not incorporate this because I was not confident, but it seems important.

Thanks...