Updates to the camera view
Have some of the widgets functioning again (e.g., channel index, min and max counts, rolling average).  Tried to get the lookup tables working again, but now that we switched to PIL I think we have to implement our own versions.  Began doing this at the bottom under calculate_LUT().

Also, the contrast adjustments are not working.  I assume something with the conversion to a a PIL Image, or perhaps the presentation of that Image by the ImageTk.PhotoImage class, is throwing that off.  Need to look into.