Addressing Dynamic Sizing #1
The app now starts out as 3/4 the size of the current display the app is running on. Can easily be changed. Went in and added the config settings for the gridded widgets. If we notice funky behavior let me know and I can make the widget sticky(should*** resize with app now).
Closes #1
For now