Software Now Runs On Mac
The software will now run on mac but will not be able to run analysis. The cross platform issues arose from Tensorflow and Cupy. So whenever those are used we will need to do a platform check. Now some development can be done from Mac. I also created a requirements_mac.txt file that can be used to create the conda env quickly.