Workaround hack...
Couldn't figure out why there was no shared memory object in Python 3.9 on my machine. Tried to instal it locally, then tried creating numpy array for the SharedNDArray and it somehow worked.  At least lets me troubleshoot remotely.

Rest of the commits hsould be minor.  Only thing that perplexed me was an error being thrown in the SyntheticZoom class.  Zoom was an array, not a string, so I had to change it to if zoom[0] in self.zoomdict...  Not sure why that happened, but I presume it will cause problems.