Cleanup
Getting a custody error.  I saw this yesterday on the Linux machine too.

stage thread exception happened! Two different threads tried to use the same ObjectInSubprocess at the same time! This is bad. Look at the docstring of concurrency_tools.py to see an example of how to use a _Custody object to avoid this problem. Traceback (most recent call last):
  File "/Users/S155475/Desktop/GitHub/ASLM/src/aslm/controller/thread_pool.py", line 146, in func
    target(*args, **kwargs)
  File "/Users/S155475/Desktop/GitHub/ASLM/src/aslm/controller/aslm_controller.py", line 754, in move_stage
    success = self.model.move_stage(pos_dict)
  File "/Users/S155475/Desktop/GitHub/ASLM/src/aslm/model/concurrency/concurrency_tools.py", line 373, in __getattr__
    with self._.parent_pipe_lock:
  File "/Users/S155475/Desktop/GitHub/ASLM/src/aslm/model/concurrency/concurrency_tools.py", line 538, in __enter__
    raise RuntimeError(
RuntimeError: Two different threads tried to use the same ObjectInSubprocess at the same time! This is bad. Look at the docstring of concurrency_tools.py to see an example of how to use a _Custody object to avoid this problem.