Solvent Selection
Added the ability to select the solvent you are imaging in within the saving popup dialog.  This is important given that it changes the physical magnification of the high-resolution module.  Need to make sure that it is also passed to the tiling wizard.  I may also eventually make a menu selection too so that it doesn't require the acquisition popup to be selected.  Not sure what is best.

Co-Authored-By: Kevin Dean <42547789+AdvancedImagingUTSW@users.noreply.github.com>