objectives
While hiding from in-laws, I thought that this could be a useful way to keep track of the different objectives that we use.  We can define which objective that each microscope is equipped with in the configuration file, then pull the properties from here or calculate some basic parameters such as the diameter of the entrance pupil.