Cleanup Autofocus
Adhered to CamelCase convention for class.  Changed throughout src and tests to tolerate change.

Commented out the toolbar, which wasn't all that useful so far.

Provided NumpyDoc for splash screen.

Want to try to address #157 today since @Elepicos put in the ground work here. One idea is to fit the DCTS response, and if the R^2 value valls below a threshold we then add alternative metrics to estimate the peak.  I also want the fit to be shown in the graphs..