Logging Work
#194

I think this may be most of the way there already.  Now the logs are being saved into LOCALAPPDATA/.ASLM/LOGS into a sub-directory that specifies the YYYY-MM-DD-HHMM.

If @codeCollision4 can look over this and make sure everything is operating as expected, I think we are good to go.  Will initiate a PR.

Co-Authored-By: Kevin Dean <42547789+AdvancedImagingUTSW@users.noreply.github.com>