Mad City
To avoid trying to connect to the device, now provide if statement...  I think this should be fine...  My logic is as follows.  If the device connection in the auto redial frameworks fails, that will throw an error.  Only when it succeeds will it be passed along.  The last thing I want though is for us to think it connected, and for it to pass a None to the MCLStage calss, which will initialize without caring...