Disable waveform button
Still not fully functional, but the goal is to be able to rapidly set the laser and galvo amplitudes to 0 volts for a quick alignment. Offset remains untouched.  I am concerned slightly that the approach isn't great. When the button is hit, it overwites the values to 0, which then I think automatically binds to update the waveform_constants instance.  Really I want to leave the original values untouched, and only quickly disable the amplitude.