Autofocus Updates
Meant to address #314

Need to make sure we are okay with how I communicated with the controller, which I know we try to minimize from the model.  At the beginning of the autofocus routine, I update the stage_controller, and at the end of the autofocus routine.

Unfortunately had to lint the controller, so I want to be absolutely sure that any formatting does not cause us problems.

I also was having problems with Ruff/Black conflicts.  My hope is that line 414 of the controller is not problematic.  Placed dictionary calls in parentheses...