Adding Sine Wave Galvo Operation
For #180 - Goal is to provide two sine waves that are 90 degrees out of phase to drive galvos for a ring TIRF application.  Ran into other errors that require further evaluation, but we have it working.  Will need to make sure that the changes to the configuration file are agreed upon by everyone, and that the proper changes are added so that we don't mess up other operation modes...

Co-Authored-By: Kevin Dean <42547789+AdvancedImagingUTSW@users.noreply.github.com>