Some odd behavior.
Doesn't appear as if the smoothing parameter is working exactly as I hoped, at least after looking at the waveform plots.

Currently, when the waveform popup is called, it loads the values from the configuration.yaml file. However, given that this is a popup window, these values are killed if you close and open the window again.   Need to perhaps pull these values from the model.