Model + filter_wheel _sutter
The filter_wheel_sutter defaulted to a hardcoded filter name, this is now removed and it goes to the 0th position.

The model has a lot of attributes that we aren't using. @annie-xd-wang, it seems like we should remove the ones we ar enot using.  I only removed 3 here (current_filter, current_laser, current_laser_index), but I wasn't brave enough to try more.