Update controller.py
Now, upon exit, the experiment.yml file will be updated. This way, whatever the GUI returns to whatever the user had selected previously. Decreases the likelihood that the user will have to annoyingly change all the parameters after rebooting the software. Of course, the user could also simply load the previous experiment.yml file, but this avoids that demand.