Update robots_mecademic.py
Robot stuff commented out, Dockable notebook appears with one robot tab and one non-robot tab. Run Entry changes colors & summons info card... Info cared is not great anymore, as it pushes robot tab of dockable notebook to the back of order or tabs, which is inconvenient. Also, there is custom widget or something that checks validity of entries... try that instead of/ in addition to color change.