Floating step size/range
Able to pass non-int values to the autofocus settings module. Works with synthetic mode. Needs to be validated.

It's not immediately clear from the presentation of the GUI if we are optimizing the remote focus amplitude or offset. If the amplitude, a negative value doesn't make sense. But I also do not necessarily see why the offset would need to go negative, since the autofocusing wizard is relative to the current value.