Adding few more menu items.
Idea is that I want to make it as easy as possible for new users to access the logs, and these files (as well as the configuration file) are somewhat hidden to the uninitiated. So, now they can call a menu item to pull it up. That way, when they have an issue, they can also submit the logs.