Update stage_sutter.py
I implemented some basic logic into the MP285 to make sure that we do not tell the stage to move, unless the move is greater than some threshold. The threshold chosen arbitrarily was 200 nm. It is specific to this class, and the MP285 should never be used to acquire image slices for a z-stack, which is the only time we would have a step that size.

Side-effect: Much faster since we are not pinging the stage at every z-slice, and this could help us avoid the stupid scrambled screen if it arises from too many pings.