Attempt at disabled widgets
Trying to change the abs_z_start, abs_z_end, and number_z_steps widgets to a disabled state. For some reason, i cannot it to work... We don't change these values, they only report on what we are doing. Thus, we should be able to avoid throwing an error if they are read only.

Could use some help.