Pull parameters from configuration file.
GalvoNIStage now looks for settle_duration_ms and distance_threshold in the configuration file.

If it is not provided, distance_threshold is set to None, and the software will not implement a wait duration.

If distance_threshold is provided, the default settle duration is 20 milliseconds.

If distance_threshold and settle duration are provided, then the software will wait the provided duration of time if the move size is larger than the distance threshold.