moved cva_ttl to cva_conpro and removved all move axis absolute in cva conpro
cva-ttl is functional multichannel external trigger. I removed all of the move axis absolute and rely on scanr to find the correct position. When the stage position is before start position to correct buffer it does send external  trigger but if the stage is too far away or after start position, system fails.