Verify that multiposition is actually true
Prior to launching the feature container with the model, we now verify that the multi position table actually has values if the multi position toggle is selected. If not, we disable the toggle in the GUI and set the multi position to False.

If this is decent, will have to think about how to put a test in...