Actual dichroic device now launches.
The shared devices section of the code is still a nightmare. Now split into three different sections, and it would really benefit from being made coherent.

On `microscope.py`, there is a shared device section. This region is by far the most confusing. I would advocate that we move away from this...

Within `device_startup_functions`, I now put a way to identify shared devices in `load_devices`, and I think this is the most clear.  You linearly move down the devices, and for any devices that can share a common controller, you do a simple check.

Lastly, in `device_startup_functions` in the `load_stages` function, there is the third option where we look for the filter wheel, and then append it to a stage device list as a string.

```            if filter_wheel == "ASI":
                stage_devices.append("shared device")
```

I advocate for option two.