Fixing some tests
This should drastically reduce the number of failures.

One problem arose in the Microscope object, where the dichroic key should be optional and only depend on the configuration of the microscope according to the yaml file.

The other problem arose where some tests required the DummyView, which now required the configuration. Now the controller is passed as a fixture.