Repair path to icon
I don't know why the relative path doesn't work anymore, but it doesn't.

To make sure that everything was actually functional, and we didn't have some problem related to the rename, I also installed a new environment on your machine (navigate), and pip installed the dev version of the pytoml.

Co-Authored-By: Kevin Dean <42547789+AdvancedImagingUTSW@users.noreply.github.com>