Still in progress.
Also does not work, but it is closer. I need to find a way to listen to the autofocus routine and only proceed turn it to live only if it is not active. To test this, I have been going into a continuous mode, then using the autofocus hotkey to trigger the autofocus routine (control + shift + a).