Update camera_hamamatsu.py
@annie-xd-wang, @Sriya-235 and I were going over some of the basics of code review, and I probably went a bit too crazy.

So, each type of Hamamatsu Camera has its own 1H/minimum exposure time value. Thus, it seems as if we should have this specified in their very own __init__ functions.

Then we were a bit confused as well. Why is it that the methods for max_image_width, which are getters and setters, are both in the parent and child classes? It seems like we could eliminate a fair amount of code if these were only defined in the parent class.  We created a new HamamatsuBase class in hopes of simplifying the children classes...  For example, I think the calculate_exposure_time is only unique for the Hamamatsu Lightning camera...