Save and load waveform constants and a few changes to waveforms.py
Hey!

So I made some changes to the menu_controller.py to save and load waveform_constants files. This allows us to  save the waveform constants for experiments and load the same constants for future experiments.

This commit also includes changes that @annie-xd-wang made to waveforms.py to align the wave.