Added devices...
Now there is automatically a way to select the devices types, which are stored in __init__.

Next I am thinking we should automatically create yet another frame after they select the device type and show two columns. One for mandatory parameters, and another for optional parameters.

There are more details lurking. For some devices, like cameras, we will need to specify the number of devices. Will move to a PR so that it isn't too big.