Fix Window Menu
The hot-keys to change windows did work, but when you manually selected the items in the menu (e.g., Channel Settings, Stage Control, etc.), it did not. Lambda function was poorly defined.

Popout camera tab was also not implemented, and now it is.

Few ruff formatting changes. Some of our calls are definitely getting too long, and this is making it harder to be compliant with pre-commit hooks.