Mainly formatting.
I'm thinking we should move our gui parameters out of the configuration.yaml file. Currently, in the channels_tab_controller, the initialize method does not call self.set_spinbox_range_limits. This is where the gui parameters are called from the configuration file. However, if the configuration file is not set up properly, then no error handling is in place and this will throw a terminal error.