Update
Updated so that it follows the general layout for most of our documentation. The troubleshooting was provided as a note, and the image of the traceback was removed. Also, given that the shutter requires a NI-based device to trigger it, and it is currently only configured to have a digital port, I have updated the documentation to specify this.