Update asi_tiger_controller.py
On John's system, there are other types of cards that are not included with our other Tiger Controller builds. This includes piezos, for example, which do not recognize some of the commands we use to set the speed as a percent of the max speed. This fix now goes through all of the stages and removes those that do not qualify.

https://www.asiimaging.com/docs/products/serial_commands#commandbuild_bu