Implementing self.data_q and getting rid of ObjectInSubprocesss
GLFrameViewer already runs the rendering on a threading.Thread. ObjectInSubprocess was reduntant and I'm surprised it worked at all. Got rid of it.

Also following the try_to_display_image method in CameraViewController using a dedicated self.data_q.

self.cmd_q is reserved only for commands which need to be sent to the shader, e.g LUT changes, etc